refactor: логика рендера новости вынесена в отдельный компонент
